requirements.txt
chess==1.11.2
google-genai==1.15.0
Jinja2==3.1.6
Flask==3.1.1
python-dotenv==1.1.0
